harmonize rpc log context
